Rename datasets file in tests
